+++ vampytest/__main__.py
@@ -2,7 +2,7 @@ import sys
 
 from scarletio import get_short_executable
 
-from .main import execute_from_terminal
+from .main import execute_from_parameters
 
 
 def __main__():
@@ -14,7 +14,8 @@ def __main__():
     if parameters and ((parameters[0] == __file__) or (parameters[0] != get_short_executable())):
         del parameters[0]
     
-    execute_from_terminal(parameters)
+    return_code = execute_from_parameters(parameters)
+    raise SystemExit(return_code)
 
 
 if __name__ == '__main__':